% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpaths.R
\name{bpaths}
\alias{bpaths}
\title{Branch Paths}
\usage{
bpaths(destination = c(i, j), branch = c(bi, bj))
}
\arguments{
\item{destination}{A vector defining the destination coordinates in terms of
a horizontal coordinate i and a vertical coordinate j}

\item{branch}{A vector defining the branch coordinates in terms of a horizontal
coordinate bi and a vertical coordinate bj}
}
\value{
A number
}
\description{
The bpaths function computes the number of paths starting at the origin (0,0)
passing through the branch coordinate (bi, bj) and ending at the destination
(i, j).  At non-branch coordinates, the path is only allowed to step upwards
and rightwards; however, at the branch coordinate, the path is allowed to step
in any direction except the direction in which the path entered the branch
coordinate.  This is similar to a no-U turn intersection where a car can go
straight, turn left, or turn right.
}
\examples{
bpaths(c(3,3), c(2,2))
}
