% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_probability_funs.R
\name{dbpaths}
\alias{dbpaths}
\title{Branch Paths Probability Mass Function}
\usage{
dbpaths(destination = c(i, j), branch = c(bi, bj))
}
\arguments{
\item{destination}{A vector defining the destination coordinates in terms of
a horizontal coordinate i and a vertical coordinate j.}

\item{branch}{A vector defining the branch location's coordinates in terms of
a horizontal coordinate bi and a vertical coordinate bj.}
}
\value{
A real number in the closed interval 0 to 1
}
\description{
A probability mass function where the number of branch paths through the
branch location specified in the branch parameter is the event of interest and
the sum of branch paths across all coordinates on the grid is the sample space.
}
\examples{
dbpaths(c(2,2), c(2,1))
}
