% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detour_probability_funs.R
\name{ddpaths}
\alias{ddpaths}
\title{Detour Paths Probability Mass Function}
\usage{
ddpaths(destination = c(i, j), detour = c(di, dj))
}
\arguments{
\item{destination}{A vector defining the destination coordinates in terms of
a horizontal coordinate i and a vertical coordinate j.}

\item{detour}{A vector defining the detour location's coordinates in terms of
a horizontal coordinate ti and a vertical coordinate tj.}
}
\value{
A real number in the closed interval 0 to 1
}
\description{
A probability mass function where the number of detour paths through the
detour location specified in the detour parameter is the event of interest and
the sum of detour paths across all coordinates on the grid is the sample space.
}
\examples{
ddpaths(c(2,2), c(2,1))
}
