% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpaths.R
\name{dpaths}
\alias{dpaths}
\title{Detour Paths}
\usage{
dpaths(destination = c(i, j), detour = c(di, dj))
}
\arguments{
\item{destination}{A vector defining the destination coordinates in terms of
a horizontal coordinate i and a vertical coordinate j}

\item{detour}{A vector defining the detour coordinates in terms of a horizontal
coordinate di and a vertical coordinate dj}
}
\value{
A number
}
\description{
The dpaths function computes the number of paths with equal Manhattan Distance
starting at the origin (0,0) NOT passing through the detour coordinate (di, dj) and
ending at the destination (i, j).
}
\examples{
dpaths(destination = c(4,4), detour = c(2,2))
}
