% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_probability_funs.R
\name{dtpaths}
\alias{dtpaths}
\title{Target Paths Probability Mass Function}
\usage{
dtpaths(destination = c(i, j), target = c(ti, tj))
}
\arguments{
\item{destination}{A vector defining the destination coordinates in terms of
a horizontal coordinate i and a vertical coordinate j.}

\item{target}{A vector defining the target location's coordinates in terms of
a horizontal coordinate ti and a vertical coordinate tj.}
}
\value{
A real number in the closed interval 0 to 1
}
\description{
A probability mass function where the number of target paths through the
target location specified in the target parameter is the event of interest and
the sum of target paths across all coordinates on the grid is the sample space.
}
\examples{
dtpaths(c(2,2), c(2,1))
}
