% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpaths.R
\name{mpaths}
\alias{mpaths}
\title{Manhattan Paths}
\usage{
mpaths(destination = c(i, j))
}
\arguments{
\item{destination}{A vector defining the destination coordinates in terms of
a horizontal coordinate i and a vertical coordinate j}
}
\value{
A number
}
\description{
The mpaths function computes the number of paths with equal Manhattan Distance
starting at the origin (0,0) and ending at (i, j).
}
\examples{
mpaths(c(2,2))
}
