% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_probability_funs.R
\name{pbpaths}
\alias{pbpaths}
\title{Branch Paths Cumulative Distribution Function}
\usage{
pbpaths(destination = c(i, j), branch = c(bi, bj))
}
\arguments{
\item{destination}{A vector defining the destination coordinates in terms of
a horizontal coordinate i and a vertical coordinate j.}

\item{branch}{A vector defining the branch coordinates in terms of a
horizontal coordinate bi and a vertical coordinate bj.}
}
\value{
A real number in the closed interval 0 to 1
}
\description{
A cumulative distribution function for the dbpaths probability mass function
}
\examples{
pbpaths(destination = c(5,5), branch = c(1,1))
}
