% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detour_probability_funs.R
\name{pdpaths}
\alias{pdpaths}
\title{Detour Paths Cumulative Distribution Function}
\usage{
pdpaths(destination = c(i, j), detour = c(di, dj))
}
\arguments{
\item{destination}{A vector defining the destination coordinates in terms of
a horizontal coordinate i and a vertical coordinate j.}

\item{detour}{A vector defining the detour coordinates in terms of a
horizontal coordinate ti and a vertical coordinate tj.}
}
\value{
A real number in the closed interval 0 to 1
}
\description{
A cumulative distribution function for the ddpaths probability mass function
}
\examples{
pdpaths(destination = c(5,5), detour = c(1,1))
}
