% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_probability_funs.R
\name{ptpaths}
\alias{ptpaths}
\title{Target Paths Cumulative Distribution Function}
\usage{
ptpaths(destination = c(i, j), target = c(ti, tj))
}
\arguments{
\item{destination}{A vector defining the destination coordinates in terms of
a horizontal coordinate i and a vertical coordinate j.}

\item{target}{A vector defining the target coordinates in terms of a
horizontal coordinate ti and a vertical coordinate tj.}
}
\value{
A real number in the closed interval 0 to 1
}
\description{
A cumulative distribution function for the dtpaths probability mass function
}
\details{
If a path passing through a coordinate represents an impression, then ptpaths
is the proportion of impressions obtained from the coordinates i <= ti &
j <= tj after one traversal of each unique Manhattan Path from the origin to
the destination.
}
\examples{
ptpaths(destination = c(5,5), target = c(1,1))
}
