% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_probability_funs.R
\name{qtpaths}
\alias{qtpaths}
\title{Target Paths Quantile Function}
\usage{
qtpaths(destination = c(i, j), p, wise = "col")
}
\arguments{
\item{destination}{A vector defining the destination coordinates in terms of
a horizontal coordinate i and a vertical coordinate j.}

\item{p}{a probability}

\item{wise}{col selects lowest coordinates by j then i while row selects lowest
by i then j}
}
\value{
coordinates for a target location
}
\description{
A quantile function which takes a probability p and a destination (i, j) as
inputs and returns the coordinates for the target location (ti, tj) where the
cumulative probability is equal to or greater than p. The target location
generated has the smallest cumulative probability >= p; if multiple coordinates
are tied, then the target location with closest Manhattan Distance to the origin
will be selected, and if there are still ties, then the wise parameter will
be used to break ties.
}
\examples{
qtpaths(c(5,5), .5)
}
