% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_probability_funs.R
\name{rbpaths}
\alias{rbpaths}
\title{Branch Paths Random Variable Function}
\usage{
rbpaths(destination = c(i, j), n = 1)
}
\arguments{
\item{destination}{A vector defining the destination coordinates in terms of
a horizontal coordinate i and a vertical coordinate j.}

\item{n}{The number of branch coordinates to generate}
}
\value{
A matrix of branch coordinates
}
\description{
A random variable function that generates random branch coordinates from the
distribution defined by the dbpaths probability mass function.
}
\examples{
rbpaths(c(10,10), 2)
}
