% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detour_probability_funs.R
\name{rdpaths}
\alias{rdpaths}
\title{Detour Paths Random Variable Function}
\usage{
rdpaths(destination = c(i, j), n = 1)
}
\arguments{
\item{destination}{A vector defining the destination coordinates in terms of
a horizontal coordinate i and a vertical coordinate j.}

\item{n}{The number of detour coordinates to generate}
}
\value{
A matrix of detour coordinates
}
\description{
A random variable function that generates random detour coordinates from the
distribution defined by the ddpaths probability mass function.
}
\examples{
rdpaths(c(10,10), 2)
}
