% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_probability_funs.R
\name{rtpaths}
\alias{rtpaths}
\title{Target Paths Random Variable Function}
\usage{
rtpaths(destination = c(i, j), n = 1)
}
\arguments{
\item{destination}{A vector defining the destination coordinates in terms of
a horizontal coordinate i and a vertical coordinate j.}

\item{n}{The number of target coordinates to generate}
}
\value{
A matrix of target coordinates
}
\description{
A random variable function that generates random target coordinates from the
distribution defined by the dtpaths probability mass function.
}
\examples{
rtpaths(c(10,10), 2)
}
