% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpaths.R
\name{tpaths}
\alias{tpaths}
\title{Target Paths}
\usage{
tpaths(destination = c(i, j), target = c(ti, tj))
}
\arguments{
\item{destination}{A vector defining the destination coordinates in terms of
a horizontal coordinate i and a vertical coordinate j}

\item{target}{A vector defining the target coordinates in terms of a horizontal
coordinate ti and a vertical coordinate tj}
}
\value{
A number
}
\description{
The tpaths function computes the number of paths with equal Manhattan Distance
starting at the origin (0,0) passing through the target coordinate (ti, tj) and
ending at the destination (i, j).
}
\examples{
tpaths(destination = c(4,4), target = c(2,2))
}
