% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpaths.R
\name{mpaths}
\alias{mpaths}
\title{Manhattan Paths}
\usage{
mpaths(i, j)
}
\arguments{
\item{i}{number of rows in the 2D lattice}

\item{j}{number of columns in the 2D lattice}
}
\value{
A number
}
\description{
The mpaths function computes the number of paths with equal Manhattan Distance
starting at the origin (0,0) and ending at (i, j).
}
\examples{
mpaths(i = 2, j = 2)
}
