% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_probability_funs.R
\name{ptpaths}
\alias{ptpaths}
\title{Target Location Probability Mass Function}
\usage{
ptpaths(destination = c(i, j), target = c(ti, tj))
}
\arguments{
\item{destination}{A vector defining the destination coordinates in terms of
a horizontal coordinate i and a vertical coordinate j.}

\item{target}{A vector defining the target coordinates in terms of a
horizontal coordinate ti and a vertical coordinate tj.}
}
\value{
A real number in the closed interval 0 to 1
}
\description{
A probability mass function where the number of Target Paths (with a target
location defined with the targer parameter) is the event of interest and the
number of Manhattan Paths (on a grid defined with the destination parameter)
is the sample space.
}
\examples{
ptpaths(destination = c(5,5), target = c(1,1))
}
