% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detour_probability_funs.R
\name{qdpaths}
\alias{qdpaths}
\title{Detour Paths Quantile Function}
\usage{
qdpaths(destination = c(i, j), p, wise = "col")
}
\arguments{
\item{destination}{A vector defining the destination coordinates in terms of
a horizontal coordinate i and a vertical coordinate j.}

\item{p}{a probability}

\item{wise}{when generating quantiles, col uses j (i.e., columns) for the
inner loop and i for the outer loop while row uses i for the inner loop
and j for the outer loop.}
}
\value{
coordinates for a detour location
}
\description{
A quantile function which takes a probability p and a destination (i, j) as
inputs and returns the coordinates for the detour location (di, dj) where the
cumulative probability is equal to or just greater than p.
}
\examples{
qdpaths(c(5,5), .5)
}
