% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_probability_funs.R
\name{qtpaths}
\alias{qtpaths}
\title{Target Location Quantile Function}
\usage{
qtpaths(destination = c(i, j), p)
}
\arguments{
\item{destination}{A vector defining the destination coordinates in terms of
a horizontal coordinate i and a vertical coordinate j.}

\item{p}{a probability}
}
\value{
coordinates for a target location
}
\description{
A quantile function which takes a probability p and a destination (i, j) as
inputs and returns the coordinates for the target location (ti, tj) where the
cumulative probability is equal to or just greater than p.
}
\examples{
qtpaths(c(5,5), .5)
}
